src/requirements.txt
# conda create -n dire python=3.9
# pip install torch==2.0.0+cu117 torchvision==0.15.1+cu117 -f https://download.pytorch.org/whl/torch_stable.html
einops
imageio
ipympl
matplotlib
numpy
opencv-python
pandas
scikit-learn
tensorboard
tensorboardX
tqdm
blobfile>=1.0.5
flask
Pillow